+++ RELEASES.md
@@ -20,6 +20,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * prov-template-compiler: PAST expression type checking (simple bindings)
 * prov-template-compiler: PAST expression type checking
 * service-templates-core: TemplateQuery: fixed issue in recursive traversal query
+* service-templates-core: fixed issues, while generating screenshots for book
 
 
 `2.2.2`